+++ server/space-service/src/main/java/edu/illinois/ncsa/incore/service/space/controllers/AllocationsController.java
@@ -10,10 +10,13 @@
 
 package edu.illinois.ncsa.incore.service.space.controllers;
 
+import edu.illinois.ncsa.incore.common.auth.IAuthorizer;
 import edu.illinois.ncsa.incore.common.dao.IGroupAllocationsRepository;
 import edu.illinois.ncsa.incore.common.dao.IUserFinalQuotaRepository;
 import edu.illinois.ncsa.incore.common.exceptions.IncoreHTTPException;
 import edu.illinois.ncsa.incore.common.utils.JsonUtils;
+import edu.illinois.ncsa.incore.common.utils.UserGroupUtils;
+import edu.illinois.ncsa.incore.common.utils.UserInfoUtils;
 import io.swagger.v3.oas.annotations.OpenAPIDefinition;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
@@ -27,6 +30,7 @@ import jakarta.ws.rs.core.Response;
 import org.apache.log4j.Logger;
 import org.json.JSONObject;
 import org.json.simple.parser.ParseException;
+import java.util.*;
 
 /**
  * Created by ywkim on 3/14/2022.
@@ -53,11 +57,27 @@ import org.json.simple.parser.ParseException;
 )
 @Path("allocations")
 public class AllocationsController {
+    private final String username;
+    private final List<String> groups;
+    private final String userGroups;
+
     @Inject
     private IGroupAllocationsRepository allocationsRepository;
 
     @Inject
     private IUserFinalQuotaRepository finalQuotaRepository;
+
+    @Inject
+    private IAuthorizer authorizer;
+
+    @Inject
+    public AllocationsController(
+        @Parameter(name = "User credentials.", required = true) @HeaderParam("x-auth-userinfo") String userInfo,
+        @Parameter(name = "User groups.", required = false) @HeaderParam("x-auth-usergroup") String userGroups) {
+        this.userGroups = userGroups;
+        this.username = UserInfoUtils.getUsername(userInfo);
+        this.groups = UserGroupUtils.getUserGroups(userGroups);
+    }
 
     private static final Logger logger = Logger.getLogger(AllocationsController.class);
 
@@ -65,10 +85,8 @@ public class AllocationsController {
     @Produces(MediaType.APPLICATION_JSON)
     @Operation(summary = "Gives the allocation and can be used as status check as well.",
         description = "This will provide the allocation of the logged in user.")
-    public String getUsage(@HeaderParam("x-auth-userinfo") String userInfo) {
+    public String getUsage() {
         JSONObject outJson = null;
-
-        String username = JsonUtils.parseUserName(userInfo);
 
         try {
             outJson = JsonUtils.createUserFinalQuotaJson(username, finalQuotaRepository);
@@ -85,15 +103,10 @@ public class AllocationsController {
     @Operation(summary = "Gives the allocation status of the given username.",
         description = "This will only work for admin user group.")
     public String getAllocationsByUsername(
-        @HeaderParam("x-auth-userinfo") String userInfo,
-        @HeaderParam("x-auth-usergroup") String userGroup,
         @Parameter(name = "Dataset Id from data service", required = true) @PathParam("username") String userId) {
-        // check if the logged in user is in admin group
-        Boolean isAdmin = JsonUtils.isLoggedInUserAdmin(userGroup);
-
         JSONObject outJson = new JSONObject();
 
-        if (isAdmin) {
+        if (this.authorizer.isUserAdmin(this.groups)) {
             try {
                 outJson = JsonUtils.createUserFinalQuotaJson(userId, finalQuotaRepository);
             } catch (ParseException e) {
@@ -114,15 +127,10 @@ public class AllocationsController {
     @Operation(summary = "Gives the allocation status of the given group name.",
         description = "This will only work for admin user group.")
     public String getAllocationsByGroupname(
-        @HeaderParam("x-auth-userinfo") String userInfo,
-        @HeaderParam("x-auth-usergroup") String userGroup,
         @Parameter(name = "Dataset Id from data service", required = true) @PathParam("groupname") String groupId) {
-        // check if the logged in user is in admin group
-        Boolean isAdmin = JsonUtils.isLoggedInUserAdmin(userGroup);
-
         JSONObject outJson = new JSONObject();
 
-        if (isAdmin) {
+        if (this.authorizer.isUserAdmin(this.groups)) {
             try {
                 outJson = JsonUtils.createGroupAllocationJson(groupId, allocationsRepository);
             } catch (ParseException e) {